Changing to int all float values  related with the stock table

+++ app/src/myanmar/java/org/eyeseetea/malariacare/presentation/factory/stock/rows/BalanceRowBuilder.java
@@ -38,7 +38,7 @@ public class BalanceRowBuilder extends StockRowBuilder {
     }
 
     @Override
-    protected Object updateColumn(Object currentValue, float newValue, SurveyStock surveyStock) {
+    protected Object updateColumn(Object currentValue, int newValue, SurveyStock surveyStock) {
         Survey survey = surveyStock.getSurvey();
         Date maxBalanceDate = Survey.getLastDateForSurveyType(Constants.SURVEY_BALANCE);
         if (survey.getType().equals(Constants.SURVEY_BALANCE) &&
